refactor: split snackbar into separate files
